Resolvers can now be specified through operationHandlers as a OperationHandlerOptions
